+++ src/main/java/joshie/harvest/core/handlers/DisableHandler.java
@@ -1,12 +1,12 @@
 package joshie.harvest.core.handlers;
 
 import joshie.harvest.core.util.HFEvents;
+import joshie.harvest.core.util.holder.HolderRegistrySet;
 import net.minecraft.entity.ai.EntityAIBase;
 import net.minecraft.entity.ai.EntityAIHarvestFarmland;
 import net.minecraft.entity.ai.EntityAITasks;
 import net.minecraft.entity.ai.EntityAITasks.EntityAITaskEntry;
 import net.minecraft.entity.passive.EntityVillager;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.event.entity.EntityJoinWorldEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
@@ -14,9 +14,7 @@ import net.minecraftforge.event.entity.player.UseHoeEvent;
 import net.minecraftforge.fml.common.eventhandler.Event.Result;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
-import java.util.HashSet;
 import java.util.Iterator;
-import java.util.Set;
 
 import static joshie.harvest.animals.HFAnimals.DISABLE_SPAWN_CHICKEN;
 import static joshie.harvest.crops.HFCrops.*;
@@ -28,16 +26,16 @@ public class DisableHandler {
     /* Disables vanilla seeds from being able to be planted **/
     @HFEvents
     public static class DisableVanillaSeeds {
-        public static final Set<Item> BLACKLIST = new HashSet<>();
+        public static final HolderRegistrySet BLACKLIST = new HolderRegistrySet();
 
         public static boolean register() {
             if (DISABLE_VANILLA_SEEDS) {
-                BLACKLIST.add(MELON_SEEDS);
-                BLACKLIST.add(PUMPKIN_SEEDS);
-                BLACKLIST.add(BEETROOT_SEEDS);
-                BLACKLIST.add(WHEAT_SEEDS);
-                BLACKLIST.add(CARROT);
-                BLACKLIST.add(POTATO);
+                BLACKLIST.register(MELON_SEEDS);
+                BLACKLIST.register(PUMPKIN_SEEDS);
+                BLACKLIST.register(BEETROOT_SEEDS);
+                BLACKLIST.register(WHEAT_SEEDS);
+                BLACKLIST.register(CARROT);
+                BLACKLIST.register(POTATO);
                 return true;
             } else return false;
         }
@@ -46,7 +44,7 @@ public class DisableHandler {
         public void onRightClickBlock(PlayerInteractEvent.RightClickBlock event) {
             ItemStack held = event.getItemStack();
             if (held != null) {
-                if (BLACKLIST.contains(held.getItem())) {
+                if (BLACKLIST.contains(held)) {
                     event.setUseItem(Result.DENY);
                 }
             }