+++ src/main/java/joshie/harvest/plugins/immersiveengineering/ImmersiveEngineering.java
@@ -1,4 +1,35 @@
 package joshie.harvest.plugins.immersiveengineering;
 
+import joshie.harvest.api.crops.Crop;
+import joshie.harvest.core.handlers.DisableHandler.DisableVanillaSeeds;
+import joshie.harvest.core.util.HFLoader;
+import net.minecraft.block.Block;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.fml.common.registry.GameRegistry.ItemStackHolder;
+
+import static joshie.harvest.api.calendar.Season.*;
+import static joshie.harvest.core.lib.HFModInfo.MODID;
+import static joshie.harvest.core.lib.LoadOrder.HFCROPS;
+
+
+@HFLoader(mods = "immersiveengineering", priority = HFCROPS + 1)
 public class ImmersiveEngineering {
+    public static final Crop HEMP = new Crop(new ResourceLocation(MODID, "hemp"), 1000L, 1L, 15, 0xB57449, SPRING, SUMMER, AUTUMN)
+                                        .setAnimalFoodType(null).setRequiresSickle(15).setGrowthHandler(new HempGrowthHandler()).setBecomesDouble(15).setRegrowStage(1);
+
+    @ItemStackHolder(value = "immersiveengineering:material", meta = 4)
+    public static final ItemStack hemp = null;
+
+    @ItemStackHolder(value = "immersiveengineering:seed")
+    public static final ItemStack hemp_seeds = null;
+
+    public static void preInit() {}
+
+    @SuppressWarnings("ConstantConditions")
+    public static void init() {
+        HEMP.setStateHandler(new StateHandlerHemp(Block.REGISTRY.getObject(new ResourceLocation("immersiveengineering", "hemp")))).setItem(hemp);
+        HEMP.setDropHandler(new DropHandlerHemp(hemp.getItem()));
+        DisableVanillaSeeds.BLACKLIST.register(hemp_seeds);
+    }
 }